requirements.txt
manim
typer
rich
openai
python-dotenv
